scripts/requirements.txt
datasets
openai
supabase
python-dotenv
